fix: GPU MSM digit overflow — use i16 for wNAF digits, re-enable in IPA
The wNAF digit type was i8 (range [-128,127]) but with the upstream
ln(n) window sizing, c can reach 9+ which produces digits in
[-256,256). This caused silent overflow and wrong MSM results in the
Dory opening proof.

Fix: use i16 for digit buffer (Zig side) and int16_t (Metal side).
Also adds identity point skip in the MSM kernel.

GPU MSM now passes Jolt verification for SHA256-2048.

+++ src/gpu/shaders/msm.metal
@@ -135,7 +135,7 @@ struct MsmParams {
 
 kernel void msm_accumulate_reduce(
     device const FpBase*  points    [[buffer(0)]],
-    device const int8_t*  digits    [[buffer(1)]],
+    device const int16_t* digits    [[buffer(1)]],
     device FpBase*        buckets   [[buffer(2)]],
     device FpBase*        results   [[buffer(3)]],
     constant MsmParams&   params    [[buffer(4)]],
@@ -164,7 +164,7 @@ kernel void msm_accumulate_reduce(
 
     // Phase 1: Bucket accumulation
     for (uint i = 0; i < num_points; i++) {
-        int8_t d = digits[digit_offset + i];
+        int16_t d = digits[digit_offset + i];
         if (d == 0) continue;
 
         FpBase px = points[i * 2];
